requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
pyjwt
python-multipart
